requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
aiofiles
alembic
bcrypt
email-validator
jinja2
passlib
pillow
python-dotenv
python-jose[cryptography]
python-multipart
qrcode
sqlalchemy
requests
pydantic-settings
# Optional (for PDF invoice rendering). On Windows requires extra native deps.
# weasyprint
